App/requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe===1.1.1
Werkzeug==1.0.1
Pillow==8.3.1
numpy==1.21.2
sagemaker==2.57.0